Artem dev

+++ src/main/java/org/example/dao/impl/DaoImpl.java
@@ -10,10 +10,12 @@ import java.util.List;
 
 public abstract class DaoImpl<T, R> implements DAO<T, R> {
 
+    public static final String GET_ALL_ENTITIES = "SELECT s FROM %s s";
     private final Class<T> clazz;
     private EntityManager em;
 
     public DaoImpl(Class<T> clazz) {
+
         this.clazz = clazz;
         this.em = getEm();
     }
@@ -21,23 +23,23 @@ public abstract class DaoImpl<T, R> implements DAO<T, R> {
     @Override
     public T create(T object) {
 
-        em.getTransaction().begin();
-        em.persist(object);
-        em.getTransaction().commit();
+        if (object != null) {
+            em.getTransaction().begin();
+            em.persist(object);
+            em.getTransaction().commit();
+        }
         return object;
     }
 
     @Override
     public T read(R id) throws EntityNotFoundException {
-        T object;
-        em.getTransaction().begin();
-        object = em.find(clazz, id);
-        em.getTransaction().commit();
-        return object;
+
+        return em.find(clazz, id);
     }
 
     @Override
     public T update(T object) {
+
         T t;
         em.getTransaction().begin();
         t = em.merge(object);
@@ -48,22 +50,19 @@ public abstract class DaoImpl<T, R> implements DAO<T, R> {
     @Override
     public void delete(R id) throws EntityNotFoundException {
 
-        em.getTransaction().begin();
-        Object rootEntity = em.getReference(clazz, id);
-        em.remove(rootEntity);
-        em.getTransaction().commit();
+        if (id != null) {
+            em.getTransaction().begin();
+            Object rootEntity = em.getReference(clazz, id);
+            em.remove(rootEntity);
+            em.getTransaction().commit();
+        }
     }
-
 
     @Override
     public List<T> readAll() {
 
-        String sqlQuery = String.format("SELECT s FROM %s s", clazz.getSimpleName());
-        em.getTransaction().begin();
-        TypedQuery<T> query = em.createQuery(sqlQuery, clazz);
-        List<T> list = query.getResultList();
-        em.getTransaction().commit();
-        return list;
+        TypedQuery<T> query = em.createQuery(getAllSqlString(), clazz);
+        return query.getResultList();
     }
 
     @Override
@@ -74,12 +73,16 @@ public abstract class DaoImpl<T, R> implements DAO<T, R> {
         }
     }
 
-
     protected EntityManager getEm() {
 
         if (em == null || !em.isOpen()) {
             em = HibernateUtil.getEntityManager();
         }
         return em;
+    }
+
+    protected String getAllSqlString() {
+
+        return String.format(GET_ALL_ENTITIES, clazz.getSimpleName());
     }
 }
